+++ GoToBible.Providers/LocalResourceOptions.cs
@@ -1,0 +1,21 @@
+﻿// -----------------------------------------------------------------------
+// <copyright file="LocalResourceOptions.cs" company="Conglomo">
+// Copyright 2020-2023 Conglomo Limited. Please see LICENSE.md for license details.
+// </copyright>
+// -----------------------------------------------------------------------
+
+namespace GoToBible.Providers;
+
+/// <summary>
+/// The Local Resource Provider Options.
+/// </summary>
+public class LocalResourceOptions
+{
+    /// <summary>
+    /// Gets or sets the Resource directory.
+    /// </summary>
+    /// <value>
+    /// The Resource directory.
+    /// </value>
+    public string ResourceDirectory { get; set; } = string.Empty;
+}
